Mac compatibility

+++ l2ork_addons/tar_em_up.sh
@@ -290,8 +290,8 @@ then
 	if [ $rpi -eq 0 ]
 	then
 		echo "installing desktop version..."
-		cp -f debian/control.desktop debian/control
-		cp -f ../../l2ork_addons/flext/config-lnx-pd-gcc.txt.intel ../../externals/grill/trunk/flext/buildsys/config-lnx-pd-gcc.txt
+		test -f debian/control.desktop && cp -f debian/control.desktop debian/control
+		test -f ../../l2ork_addons/flext/config-lnx-pd-gcc.txt.intel && cp -f ../../l2ork_addons/flext/config-lnx-pd-gcc.txt.intel ../../externals/grill/trunk/flext/buildsys/config-lnx-pd-gcc.txt
 	else
 		echo "installing raspbian version..."
 		cp -f debian/control.raspbian debian/control
